Rename utils to csv and move in lib folder
