+++ src/components/Header.js
@@ -26,9 +26,5 @@ const StyledHeader = styled.header`
     display: flex;
     align-items: center;
   }
-
-  a:visited {
-    color: var(--grey);
-  }
 `
